A bit of package cleanup and reorganization in packageloader
